% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_rmst.R
\name{adj_rmst}
\alias{adj_rmst}
\title{Adjusted RMST}
\usage{
adj_rmst(
  time,
  status,
  exposure,
  data,
  ps = NULL,
  stabilize = TRUE,
  tau = NULL,
  var.est = FALSE,
  alpha = 0.05
)
}
\arguments{
\item{time}{The name of the variable for time to event}

\item{status}{The name of the variable for status (0 if censored, 1 if event)}

\item{exposure}{The name of the variable for exposure (0 if unexposed, 1 if exposed)}

\item{data}{A data frame in which contains the follow-up time (time), the event (status), the exposure (exposure), and the optional propensity score (ps)}

\item{ps}{The name of the variable for propensity score variable P(A=1|L)>0, Default: NULL}

\item{stabilize}{A logical value. If true, stabilized ipw., Default: TRUE}

\item{tau}{pre-specified time point to estimate of restricted mean survival time (RMST), Default: NULL}

\item{var.est}{A logical value. variance of RMST via xie (2005) is calculated., Default: FALSE}

\item{alpha}{A significance level, Default: 0.05}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Function for RMST estimate from adjusted Kaplan-Meier curve via SIPW (or IPW)
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 # EXAMPLE1
 library(survival)

 dat <- gbsg
 dat$size2 <- ifelse(dat$size <= 20, 0,
                     ifelse(dat$size > 20 & dat$size <= 50, 1, 2))
 dat$age2 <- dat$age/100
 dat$er2 <- dat$er/1000

 ## Estimation of propensity score
 denom.fit <- glm(hormon ~ (age2)^3 + (age2)^3*log(age2) + meno + factor(size2) + sqrt(nodes) + er2,
                  data=dat, family=binomial(link='logit'))
 dat$Ps <- predict(denom.fit, type='response')

 ## Adjusted RMST with tau qual to 5-year
 rr <- adj_rmst(data=dat, time.v='rfstime', status.v='status', exposure.v='hormon',
                ps.v='Ps', stabilize=FALSE, tau=365.25*5, alpha=.05, var.est = TRUE)
 round(c(rr$rmst_diff, rr$rmst_diff_low, rr$rmst_diff_upp, rr$rmst_diff_pval), 3)
 }
}
}
