% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{autoplot}
\alias{autoplot}
\title{Create a complete ggplot appropriate to a particular data type}
\description{
This object is imported from other package (\bold{ggplot2}). Follow the link below to see its documentation.

\describe{
\item{ggplot2}{\code{\link[ggplot2]{autoplot}}}
}
}
