% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.RMSTsensitivity.R
\name{print.RMSTsensitivity}
\alias{print.RMSTsensitivity}
\title{Print for "senRMST" objects}
\usage{
\method{print}{RMSTsensitivity}(x, digits = max(1L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class "senRMST"}

\item{digits}{print digits, Default: max(1L, getOption("digits") - 3L)}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Print for objects of class "senRMST".
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
