% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMST_sensitivity.R
\name{RMST_sensitivity}
\alias{RMST_sensitivity}
\title{Sensitivity ananlysis for RMST}
\usage{
RMST_sensitivity(
  time,
  status,
  exposure,
  exposed.ref.level = 1,
  ps,
  data,
  methods = "Approx",
  use.multicore = TRUE,
  n.core = parallel::detectCores()/2,
  lambda = 2,
  tau = NULL,
  ini.par = 1,
  verbose = FALSE
)
}
\arguments{
\item{time}{The name of the variable for time to event}

\item{status}{The name of the variable for status (0 if censored, 1 if event)}

\item{exposure}{The name of the variable for exposure (0 if unexposed, 1 if exposed)}

\item{exposed.ref.level}{Reference level in exposure variable, Default: 1}

\item{ps}{The name of the variable for propensity score variable i.e., P(A=1|L)>0}

\item{data}{A data frame in which contains the follow-up time (time), the event (status), the exposure (exposure), and the propensity score (ps)}

\item{methods}{A character with the methods how to calculate the shifted RMST ("Optim", "Approx", "LP1", "LP2"), Default: 'Approx'}

\item{use.multicore}{PARAM_DESCRIPTION, Default: TRUE}

\item{n.core}{PARAM_DESCRIPTION, Default: parallel::detectCores()/2}

\item{lambda}{PARAM_DESCRIPTION, Default: 1.5}

\item{tau}{PARAM_DESCRIPTION, Default: NULL}

\item{ini.par}{PARAM_DESCRIPTION, Default: 1}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Function for sensitivity analysis of unmeasured confounding for restricted mean survival time
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[parallel]{detectCores}}, \code{\link[parallel]{makeCluster}}
}
