% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSTSens-package.R
\docType{package}
\name{RMSTSens-package}
\alias{RMSTSens-package}
\title{RMSTSens: Sensitivity analysis for unmeasured confounding in estimating the difference in restricted mean survival time}
\description{
R package \bold{RMSTSens} is sensitivity analysis package for the difference in adjusted restricted mean survival time (RMST)
to possible unmeasured confounding
}
\references{
Lee S, Park JH, Lee W (2023):
Sensitivity analysis for unmeasured confounding in estimating the difference in restricted mean survival time
\emph{xxx},
DOI: xxx
}
\author{
Seungjae Lee \email{seungjae2525@gmail.com} and Woojoo Lee \email{lwj221@gmail.com}
}
\keyword{package}
