% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSTSens.ci.R
\name{RMSTSens.ci}
\alias{RMSTSens.ci}
\title{Confidence interval for population sensitivity range}
\usage{
RMSTSens.ci(
  x,
  B = 1000,
  level = 0.95,
  seed = NULL,
  formula,
  model = "logistic",
  trunc.prop = 0,
  use.multicore = TRUE,
  n.core = parallel::detectCores()/2,
  ...
)
}
\arguments{
\item{x}{An object for class \code{RMSTSens}. If you want to input several \code{RMSTSens} objects,
use the \code{RMSTSens.merge} function. See \code{\link[RMSTSens]{RMSTSens.merge}}.}

\item{B}{The number of bootstrap replicates. Default: 1000.}

\item{level}{The confidence level required (i.e., \eqn{1-\alpha}). Default: 0.95.}

\item{seed}{The seed number. If the propensity score was estimated using methods in the \code{randomForest} or \code{gbm} package,
then enter the seed number used at that time.}

\item{formula}{The formula for estimating propensity score. See Examples.}

\item{model}{The method for estimating propensity score. "logistic", "rf", or "gbm" can be available.
If model is not "logistic", \code{randomForest} or \code{gbm} package is needed. Default: "logistic"}

\item{trunc.prop}{Optional truncation percentile (Values between 0 and 0.5 are allowed).
For example, when trunc.prop=0.01, the left tail is truncated to the 1st percentile,
and the right tail is truncated to the 99th percentile. When specified, truncated propensity scores are returned. Default: 0.}

\item{use.multicore}{Logical scalar indicating whether to parallelize our optimization problem. Default: TRUE.}

\item{n.core}{The number of CPU cores to use. Default: parallel::detectCores()/2.}

\item{\dots}{Additional arguments passed on to \code{\link[randomForest]{randomForest}} or \code{\link[gbm]{gbm}} function in \code{randomForest} or \code{gbm} package.}
}
\value{
The object is a data.frame with class \code{RMSTSens}. The function returns following components:
\item{N}{Total number of subjects}
\item{N.exposed}{The number of subjects in exposed group}
\item{N.unexposed}{The number of subjects in unexposed group}
\item{N.event.exposed}{The number of events in exposed group}
\item{N.event.unexposed}{The number of events in unexposed group}
\item{cen.rate}{Total censoring rate}
\item{cen.rate.exposed}{Censoring rate in exposed group}
\item{cen.rate.unexposed}{Censoring rate in unexposed group}
\item{Lambda}{A scalar or vector of sensitivity parameter(s) \eqn{\Lambda} used}
\item{Tau}{User-specific time point \eqn{\tau}; If tau not specified (NULL), the minimum value of last event times in each group}
\item{Method}{A optimization method used}
\item{min.exposed}{The minimum value of adjusted RMST for exposed group}
\item{max.exposed}{The maximum value of adjusted RMST for exposed group}
\item{min.unexposed}{The minimum value of adjusted RMST for unexposed group}
\item{max.unexposed}{The maximum value of adjusted RMST for unexposed group}
\item{RMST.diff.min}{Lower bound of the sensitivity range for the difference in adjusted RMST}
\item{RMST.diff.max}{Upper bound of the sensitivity range for the difference in adjusted RMST}
\item{min.exposed.lower}{Lower (\eqn{\alpha/2})-quantile of adjusted RMST for exposed group}
\item{max.exposed.upper}{Upper (\eqn{\alpha/2})-quantile of adjusted RMST for exposed group}
\item{min.unexposed.lower}{Lower (\eqn{\alpha/2})-quantile of adjusted RMST for unexposed group}
\item{max.unexposed.upper}{Upper (\eqn{\alpha/2})-quantile of adjusted RMST for unexposed group}
\item{RMST.diff.min.lower}{Lower bound of (\eqn{1-\alpha})\% percentile bootstrap confidence interval for the population sensitivity range}
\item{RMST.diff.max.upper}{Upper bound of (\eqn{1-\alpha})\% percentile bootstrap confidence interval for the population sensitivity range}
The results for the \code{\link{RMSTSens.ci}} are printed with the \code{\link{print.RMSTSens}} function.
To generate the plot of results for the \code{\link{RMSTSens.ci}}, use the \code{\link{autoplot.RMSTSens}} function.
}
\description{
\code{RMSTSens.ci()} is the main function of \code{RMSTSens} package and
constructs the percentile bootstrap confidence interval(s) for population sensitivity range.
}
\details{
To assess details of method for sensitivity analysis, see Lee et al. (2024).
When estimating the sensitivity range of the difference in adjusted RMST using \code{\link{RMSTSens}} function,
there is no need for the argument of formula and model to estimate the propensity score.
However, when estimating the confidence interval for the population sensitivity range, the argument of formula and model are absolutely necessary.
Note that to use \code{\link{RMSTSens.ci}} function in this package, propensity score should be re-estimated by using
\enumerate{
\item \code{\link[stats]{glm}} that has a binomial distribution with logit link function or
\item \code{\link[randomForest]{randomForest}} or \code{\link[gbm]{gbm}} function in \code{randomForest} or \code{gbm} package.
}

Also, if propensity score was re-estimated using method either \code{randomForest} or \code{gbm} package,
then you should enter the seed number used at that time in "seed" argument.
}
\examples{
dat <- gbsg

## Estimation of propensity score
denom.fit <- glm(hormon~age+meno+size+factor(grade)+nodes+pgr+er,
                 data=dat, family=binomial(link="logit"))
dat$Ps <- predict(denom.fit, type="response")

## Performing the sensitivity analysis - sensitivity range
# Using approximate optimization method
results.approx2 <- RMSTSens(time="rfstime", status="status", exposure="hormon",
                            level.exposed="1", ps="Ps", data=dat, methods="Approx",
                            use.multicore=TRUE, n.core=2,
                            lambda=c(1,1.1), tau=365.25*5, ini.par=1, verbose=FALSE)

# Additional sensitivity analysis when lambda=1.2
results.approx3 <- RMSTSens(time="rfstime", status="status", exposure="hormon",
                            level.exposed="1", ps="Ps", data=dat, methods="Approx",
                            use.multicore=TRUE, n.core=2,
                            lambda=c(1.2), tau=365.25*5, ini.par=1, verbose=FALSE)

# Percentile bootstrap CI for population sensitivity range
re.ap.boot <- RMSTSens.ci(x=RMSTSens.merge(x=list(results.approx2, results.approx3)),
              B=50, # Set B=50 to reduce computation time for R checks
              level=0.95, seed=220524,
              formula=hormon~age+meno+size+factor(grade)+nodes+pgr+er,
              model="logistic", trunc.prop=0, use.multicore=TRUE, n.core=2)
re.ap.boot

}
\references{
{
Lee, S., Park, J. H., and Lee, W.
Sensitivity analysis for unmeasured confounding in estimating the difference in restricted mean survival time.
\emph{Statistical Methods in Medical Research}. 2024.
\doi{10.1177/09622802241280782}

Zhao, Q., Small, D. S., and Bhattacharya, B. B.
Sensitivity analysis for inverse probability weighting estimators via the percentile bootstrap.
\emph{Journal of the Royal Statistical Society Series B: Statistical Methodology}. 2019.
\doi{10.1111/rssb.12327}
}
}
\seealso{
\code{\link[RMSTSens]{RMSTSens}}, \code{\link[RMSTSens]{print.RMSTSens}}, \code{\link[RMSTSens]{autoplot.RMSTSens}}
}
\keyword{methods}
