% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSTSens.merge.R
\name{RMSTSens.merge}
\alias{RMSTSens.merge}
\title{Merging \code{RMSTSens} objects for the results of sensitivity analysis}
\usage{
RMSTSens.merge(x = list())
}
\arguments{
\item{x}{A list of objects for class \code{RMSTSens}. For example, list(result1, result2, ...). See Examples.}
}
\value{
New merged object for class \code{RMSTSens}
}
\description{
\code{RMSTSens} is the function to merge the \code{RMSTSens} object, which includes the result of the previously performed sensitivity analysis,
and the additional \code{RMSTSens} object, which includes the result of the another performed sensitivity analysis.
}
\examples{
dat <- gbsg

## Estimation of propensity score
denom.fit <- glm(hormon~age+meno+size+factor(grade)+nodes+pgr+er,
                 data=dat, family=binomial(link="logit"))
dat$Ps <- predict(denom.fit, type="response")

## Performing the sensitivity analysis - sensitivity range
# Using approximate optimization method
results.approx2 <- RMSTSens(time="rfstime", status="status", exposure="hormon",
                            level.exposed="1", ps="Ps", data=dat, methods="Approx",
                            use.multicore=TRUE, n.core=2,
                            lambda=c(1,1.1,1.2), tau=365.25*5, ini.par=1, verbose=FALSE)
RMSTSens.merge(x=list(results.approx2))

# Additional sensitivity analysis when lambda=1.15
results.approx3 <- RMSTSens(time="rfstime", status="status", exposure="hormon",
                            level.exposed="1", ps="Ps", data=dat, methods="Approx",
                            use.multicore=TRUE, n.core=2,
                            lambda=c(1.15), tau=365.25*5, ini.par=1, verbose=FALSE)
RMSTSens.merge(x=list(results.approx2, results.approx3))

}
