% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSTSens.R
\name{RMSTSens}
\alias{RMSTSens}
\title{Sensitivity range of the difference in adjusted RMST}
\usage{
RMSTSens(
  time,
  status,
  exposure,
  level.exposed = "1",
  ps,
  data,
  methods = "Approx",
  use.multicore = TRUE,
  n.core = parallel::detectCores()/2,
  lambda = 2,
  tau = NULL,
  ini.par = 1,
  verbose = FALSE
)
}
\arguments{
\item{time}{The name of variable for survival time (i.e., time to event).}

\item{status}{The name of variable for event indicator (0 if censored, 1 if event).}

\item{exposure}{The name of variable for exposure (0 if unexposed, 1 if exposed).}

\item{level.exposed}{Level for exposed group in exposure variable. Default: "1".}

\item{ps}{The name of variable or the vector for the estimated propensity score.}

\item{data}{A data frame in which contains the follow-up time (time), the event (status), the exposure (exposure), and the propensity score (ps).}

\item{methods}{A character with the methods how to calculate the adjusted RMST ("Optim", "Approx", "LP1", "LP2"). Default: "Approx". See Details.}

\item{use.multicore}{Logical scalar indicating whether to parallelize our optimization problem. Default: TRUE.}

\item{n.core}{The number of CPU cores to use. Default: parallel::detectCores()/2.}

\item{lambda}{A scalar or vector for sensitivity parameter \eqn{\Lambda}. Default: 2.}

\item{tau}{User-specific time point, If tau not specified (i.e., NULL), use the minimum value of last event times in each group. Default: NULL.}

\item{ini.par}{Initial value for the parameter to be optimized over in the direct optimization method. Default: 1.}

\item{verbose}{Conditional on the verbose level, print the message when each optimization (minimization or maximization) for each group was ended. Default: FALSE.}
}
\value{
An object of class \code{RMSTSens}. The object is a data.frame with the following components:
\item{N}{Total number of subjects}
\item{N.exposed}{The number of subjects in exposed group}
\item{N.unexposed}{The number of subjects in unexposed group}
\item{N.event.exposed}{The number of events in exposed group}
\item{N.event.unexposed}{The number of events in unexposed group}
\item{cen.rate}{Total censoring rate}
\item{cen.rate.exposed}{Censoring rate in exposed group}
\item{cen.rate.unexposed}{Censoring rate in unexposed group}
\item{Lambda}{A scalar or vector of sensitivity parameter(s) \eqn{\Lambda} used}
\item{Tau}{User-specific time point \eqn{\tau}, If tau not specified (NULL), use the minimum value of last event times in each group}
\item{Method}{A optimization method used}
\item{min.exposed}{The minimum value of adjusted RMST for exposed group}
\item{max.exposed}{The maximum value of adjusted RMST for exposed group}
\item{min.unexposed}{The minimum value of adjusted RMST for unexposed group}
\item{max.unexposed}{The maximum value of adjusted RMST for unexposed group}
\item{RMST.diff.min}{Lower bound of the sensitivity range for the difference in adjusted RMST}
\item{RMST.diff.max}{Upper bound of the sensitivity range for the difference in adjusted RMST}
The results for the \code{RMSTSens} are printed with the \code{\link{print.RMSTSens}} functions.
To generate the plot of results for the \code{RMSTSens}, use the \code{\link{autoplot.RMSTSens}} functions.
}
\description{
\code{RMSTSens()} is the main function of RMSTSens and
performs the sensitivity analysis for estimating the sensitivity range(s) of the difference in adjusted RMST
}
\details{
There are four possible methods for our sensitivity analysis.

In general settings,
\itemize{
\item methods="Approx": In general survival analysis setting, if the censoring rate is less than 0.7,
the approximate optimization method can be recommended
because it is much faster than and very accurate as the direct optimization method.
\item methods="Optim": If the censoring rate is greater than 0.7, the direct optimization method can be used as an alternative
because it is implemented as fast as the approximate optimization method.
}

In special settings, some analytic results can be obtained.
\itemize{
\item methods="LP1": When a closed cohort where all subjects are followed up from the same entry time and
only administrative censoring is allowed at the end of follow-up is considered,
high-dimensional optimization problems can be expressed as the well-known linear programming problems,
and thus one can use the analytic solutions for computing the sensitivity range.
\item methods="LP2": Similarly, when the minimum value of censoring times in each group is longer than or equal to tau,
the optimization problems are also transformed to well-known linear programming problems,
and thus one can use the analytic solutions for computing the sensitivity range.
}

See Lee et al. (2023) for details.
}
\examples{
dat <- gbsg
dat$size2 <- ifelse(dat$size <= 20, 0,
                    ifelse(dat$size > 20 & dat$size <= 50, 1, 2))
dat$age2 <- dat$age/100
dat$er2 <- dat$er/1000

## Estimation of propensity score
denom.fit <- glm(hormon~(age2)^3+(age2)^3*log(age2)+meno+factor(size2)+sqrt(nodes)+er2,
                 data=dat, family=binomial(link="logit"))
dat$Ps <- predict(denom.fit, type="response")

## Between-group difference in adjusted RMST based on shifted propensity score
## Adjusted RMST with tau equal to 5-year
# Using direct optimization method
results.optim <- RMSTSens(time="rfstime", status="status", exposure="hormon",
                          level.exposed="1", ps="Ps", data=dat, methods="Optim",
                          use.multicore=TRUE, n.core=2,
                          lambda=1.5, tau=365.25*5, ini.par=1, verbose=FALSE)
results.optim

# Using approximate optimization method
results.approx <- RMSTSens(time="rfstime", status="status", exposure="hormon",
                           level.exposed="1", ps="Ps", data=dat, methods="Approx",
                           use.multicore=TRUE, n.core=2,
                           lambda=1.5, tau=365.25*5, ini.par=1, verbose=FALSE)
results.approx

## Adjusted RMST with not specified tau and with multiple lambda
# Using approximate optimization method
results.approx2 <- RMSTSens(time="rfstime", status="status", exposure="hormon",
                            level.exposed="1", ps="Ps", data=dat, methods="Approx",
                            use.multicore=TRUE, n.core=2,
                            lambda=c(1,1.5), tau=365.25*5, ini.par=1, verbose=FALSE)
results.approx2

}
\references{
Lee S, Park JH, Lee W (2023):
Sensitivity analysis for unmeasured confounding in estimating the difference in restricted mean survival time.
\emph{xxx}. DOI: xxx.
}
\seealso{
\code{\link[RMSTSens]{print.RMSTSens}}, \code{\link[RMSTSens]{autoplot.RMSTSens}}, \code{\link[RMSTSens]{RMSTSens.ci}}
}
\keyword{methods}
