% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensRMST-package.R
\docType{package}
\name{sensRMST-package}
\alias{sensRMST}
\alias{sensRMST-package}
\title{sensRMST: Sensitivity Analysis for Restricted Mean Survival Time}
\description{
Sensitivity analysis of unmeasured confounding for restricted mean survival time based on propensity score.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/seungjae2525/sensRMST}
  \item Report bugs at \url{https://github.com/seungjae2525/sensRMST/issues}
}

}
\author{
\strong{Maintainer}: Seungjae Lee \email{seungjae2525@gmail.com} (\href{https://orcid.org/0000-0001-5508-8634}{ORCID})

Authors:
\itemize{
  \item Woojoo Lee (\href{https://orcid.org/0000-0001-7447-7045}{ORCID})
}

}
\keyword{internal}
