% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSTSens-package.R
\name{autoplot}
\alias{autoplot}
\title{bigobenchmark exported operators and S3 methods}
\description{
The following functions are imported and then re-exported from the RMSTSens package to avoid loading them.
}
