% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{autoplot}
\alias{autoplot}
\title{Objects exported from other package}
\description{
This object is imported from other package. Follow the links below to see its documentation.

\describe{
\item{ggplot2}{\code{\link[ggplot2]{autoplot}}}
}
}
