% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.RMSTSens.R
\name{print.RMSTSens}
\alias{print.RMSTSens}
\title{Print for "RMSTSens" objects}
\usage{
\method{print}{RMSTSens}(x, digits = max(1L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class "RMSTSens"}

\item{digits}{print digits, Default: max(1L, getOption("digits") - 3L)}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print for objects of class "RMSTSens".
}
\details{
print.RMSTSens tries to be smart about results for sensitivity analysis
}
\examples{
if(interactive()){
 dat <- gbsg
 dat$size2 <- ifelse(dat$size <= 20, 0,
                     ifelse(dat$size > 20 & dat$size <= 50, 1, 2))
 dat$age2 <- dat$age/100
 dat$er2 <- dat$er/1000

 ## Estimation of propensity score
 denom.fit <- glm(hormon~(age2)^3+(age2)^3*log(age2)+meno+factor(size2)+sqrt(nodes)+er2,
                  data=dat, family=binomial(link='logit'))
 dat$Ps <- predict(denom.fit, type='response')

 ## Between-group difference in adjusted RMST based on shifted propensity score
 ## Adjusted RMST with tau equal to 5-year
 # Using direct optimization method
 results.optim <- RMSTsensitivity(time='rfstime', status='status', exposure='hormon',
                                  exposed.ref.level=1, ps='Ps', data=dat, methods='Optim',
                                  use.multicore=TRUE, n.core=parallel::detectCores()/2,
                                  lambda=1.5, tau=365.25*5, ini.par=1, verbose=FALSE)
 print(results.optim)

 # Using approximate optimization method
 results.approx <- RMSTsensitivity(time='rfstime', status='status', exposure='hormon',
                                   exposed.ref.level=1, ps='Ps' ,data=dat, methods='Approx',
                                   use.multicore=TRUE, n.core=2,
                                   lambda=1.5, tau=365.25*5, ini.par=1, verbose=FALSE)
 print(results.approx)

 ## Adjusted RMST with not specified tau and with multiple lambda
 # Using approximate optimization method
 results.approx2 <- RMSTsensitivity(time='rfstime', status='status', exposure='hormon',
                                    exposed.ref.level=1, ps='Ps' ,data=dat, methods='Approx',
                                    use.multicore=TRUE, n.core=2,
                                    lambda=c(1,1.5), tau=365.25*5, ini.par=1, verbose=FALSE)
 print(results.approx2)

 re.ap.boot <- boot.ci.RMST(x=results.approx2, B=20, level=0.95, seed=220524,
               formula=hormon~(age2)^3+(age2)^3*log(age2)+meno+factor(size2)+sqrt(nodes)+er2,
               model="logistic", use.multicore=TRUE, n.core=2, verbose=TRUE)
 print(re.ap.boot)
}

}
