% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSTSens-package.R
\docType{package}
\name{RMSTSens-package}
\alias{RMSTSens-package}
\title{RMSTSens: Sensitivity analysis for unmeasured confounding in estimating the difference in restricted mean survival time}
\description{
R package \bold{RMSTSens} is sensitivity analysis package for the difference in adjusted restricted mean survival time (RMST)
to possible unmeasured confounding.
}
\references{
Lee, S., Park, J. H., and Lee, W.
Sensitivity analysis for unmeasured confounding in estimating the difference in restricted mean survival time.
\emph{Statistical Methods in Medical Research}. 2024.
\doi{10.1177/09622802241280782}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/seungjae2525/RMSTSens}
  \item Report bugs at \url{https://github.com/seungjae2525/RMSTSens/issues}
}

}
\author{
Seungjae Lee \email{seungjae2525@gmail.com} and Woojoo Lee \email{lwj221@gmail.com}
}
\keyword{package}
