% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSTSens-package.R
\docType{package}
\name{RMSTSens-package}
\alias{RMSTSens-package}
\alias{RMSTSens}
\title{RMSTSens: Sensitivity analysis for RMST}
\description{
R package \bold{RMSTSens} is sensitivity analysis package for RMST
}
\references{
Bakbergenuly I, Hoaglin DC, Kulinskaya E (2020):
Methods for estimating between-study variance and overall
effect in meta-analysis of odds-ratios.
\emph{Research Synthesis Methods},
DOI: 10.1002/jrsm.1404
}
\author{
Seungjae Lee \email{seungjae2525@gmail.com}
}
\keyword{package}
