% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.RMSTSens.R
\name{RMSTSens.ci}
\alias{RMSTSens.ci}
\title{Confidence interval for bias-adjusted RMST}
\usage{
RMSTSens.ci(
  x,
  B = 1000,
  level = 0.95,
  seed = 920818,
  formula,
  model = "logistic",
  use.multicore = TRUE,
  n.core = parallel::detectCores()/2,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{An object of class \code{RMSTSens}.}

\item{B}{The number of bootstrap replicates, Default: 1000.}

\item{level}{The confidence level required (i.e., \eqn{1-\alpha}), Default: 0.95.}

\item{seed}{The seed number. If the propensity score was estimated using methods in the \code{caret} package, then should enter the seed number used at that time.}

\item{formula}{The formula for estimating propensity score.}

\item{model}{The method for estimating propensity score, Default: "logistic".}

\item{use.multicore}{Logical scalar indicating whether to parallelize our optimization problem, Default: TRUE.}

\item{n.core}{The number of cores to use, Default: parallel::detectCores()/2.}

\item{verbose}{According to the verbose level, whether or not print the completion message for each 100th bootstrap, Default: TRUE.}

\item{\dots}{Additional arguments passed on to \code{train} function in \code{caret} package.}
}
\value{
The object is a data.frame with class \code{RMSTSens}. The function returns following components:
\item{N}{Total number of subjects}
\item{N.exposed}{The number of subjects in exposed group}
\item{N.unexposed}{The number of subjects in unexposed group}
\item{N.event.exposed}{The number of events in exposed group}
\item{N.event.unexposed}{The number of events in unexposed group}
\item{cen.rate}{Total censoring rate}
\item{cen.rate.exposed}{Censoring rate in exposed group}
\item{cen.rate.unexposed}{Censoring rate in unexposed group}
\item{Lambda}{A scalar or vector of sensitivity parameter \eqn{Lambda} used}
\item{Tau}{User-specific time point \eqn{tau}, If tau not specified (NULL), use the minimum between the largest observed event time in each groups}
\item{Method}{A optimization method used}
\item{min.exposed}{The minimum of adjusted RMST based on the shifted propensity score for exposed group}
\item{max.exposed}{The maximum of adjusted RMST based on the shifted propensity score for exposed group}
\item{min.unexposed}{The minimum of adjusted RMST based on the shifted propensity score for unexposed group}
\item{max.unexposed}{The maximum of adjusted RMST based on the shifted propensity score for unexposed group}
\item{RMST.diff.min}{Lower bound of point estimate for between-group difference in adjusted RMST based on shifted propensity score}
\item{RMST.diff.max}{Upper bound of point estimate for between-group difference in adjusted RMST based on shifted propensity score}
\item{RMST.diff.min.lower}{Lower bound of (\eqn{1-\alpha})\% confidence interval for between-group difference in adjusted RMST based on shifted propensity score}
\item{RMST.diff.max.upper}{Upper bound of (\eqn{1-\alpha})\% confidence interval for between-group difference in adjusted RMST based on shifted propensity score}
\item{min.exposed.lower}{Lower (\eqn{\alpha/2})-quantile of adjusted RMST based on the shifted propensity score for exposed group}
\item{max.exposed.upper}{Upper (\eqn{\alpha/2})-quantile of adjusted RMST based on the shifted propensity score for exposed group}
\item{min.unexposd.lower}{Lower (\eqn{\alpha/2})-quantile of adjusted RMST based on the shifted propensity score for unexposed group}
\item{max.unexposed.upper}{Upper (\eqn{\alpha/2})-quantile of adjusted RMST based on the shifted propensity score for unexposed group}
The results for the \code{\link{RMSTSens.ci}} are printed with the \code{\link{print.RMSTSens}} functions.
To generate results plot comparing Lambda with confidence interval and range of adjusted RMST based on shifted propensity score, use the \code{\link{plot.RMSTSens}} function.
}
\description{
Function for constructing the confidence interval for restricted mean survival time using propensity score
}
\details{
To assess details of method for sensitivity analysis, see Lee et al. (2022) for details.
}
\examples{
if(interactive()){
 dat <- gbsg
 dat$size2 <- ifelse(dat$size <= 20, 0,
                     ifelse(dat$size > 20 & dat$size <= 50, 1, 2))
 dat$age2 <- dat$age/100
 dat$er2 <- dat$er/1000

 ## Estimation of propensity score
 denom.fit <- glm(hormon~(age2)^3+(age2)^3*log(age2)+meno+factor(size2)+sqrt(nodes)+er2,
                  data=dat, family=binomial(link='logit'))
 dat$Ps <- predict(denom.fit, type='response')

 ## Between-group difference in adjusted RMST based on shifted propensity score
 ## Adjusted RMST with not specified tau and with multiple lambda
 # Using approximate optimization method
 results.approx2 <- RMSTSens(time='rfstime', status='status', exposure='hormon',
                             exposed.ref.level=1, ps='Ps' ,data=dat, methods="Approx",
                             use.multicore=TRUE, n.core=2,
                             lambda=c(1,1.5), tau=365.25*5, ini.par=1, verbose=FALSE)
 re.ap.boot <- RMSTSens.ci(x=results.approx2, B=20, level=0.95, seed=220524,
               formula=hormon~(age2)^3+(age2)^3*log(age2)+meno+factor(size2)+sqrt(nodes)+er2,
               model="logistic", use.multicore=TRUE, n.core=2, verbose=TRUE)
 re.ap.boot
}

}
\references{
Bakbergenuly I, Hoaglin DC, Kulinskaya E (2020):
Methods for estimating between-study variance and overall
effect in meta-analysis of odds-ratios.
\emph{Research Synthesis Methods},
DOI: 10.1002/jrsm.1404
}
\seealso{
\code{\link[RMSTSens]{print.RMSTSens}}, \code{\link[RMSTSens]{plot.RMSTSens}}
}
\author{
Seungjae Lee \email{seungjae2525@gmail.com}
}
\keyword{methods}
