% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSTSens.R
\name{RMSTSens}
\alias{RMSTSens}
\title{Sensitivity analysis for RMST}
\usage{
RMSTSens(
  time,
  status,
  exposure,
  level.exposed = "1",
  ps,
  data,
  methods = "Approx",
  use.multicore = TRUE,
  n.core = parallel::detectCores()/2,
  lambda = 2,
  tau = NULL,
  ini.par = 1,
  verbose = FALSE
)
}
\arguments{
\item{time}{The name of variable for time to event.}

\item{status}{The name of variable for censoring indicator (0 if censored, 1 if event).}

\item{exposure}{The name of variable for exposure (0 if unexposed, 1 if exposed).}

\item{level.exposed}{Level for exposed group in exposure variable, Default: "1".}

\item{ps}{The name of variable for propensity score variable or the vector for propensity score i.e., P(A=1|L).}

\item{data}{A data frame in which contains the follow-up time (time), the event (status), the exposure (exposure), and the propensity score (ps).}

\item{methods}{A character with the methods how to calculate the adjusted RMST ("Optim", "Approx", "LP1", "LP2"), Default: "Approx". See Details.}

\item{use.multicore}{Logical scalar indicating whether to parallelize our optimization problem, Default: TRUE.}

\item{n.core}{The number of cores to use, Default: parallel::detectCores()/2.}

\item{lambda}{A scalar or vector for sensitivity parameter \eqn{\Lambda}, Default: 2.}

\item{tau}{User-specific time point, If tau not specified (NULL), use the minimum of the largest observed event time in both groups, Default: NULL.}

\item{ini.par}{Initial parameter for direct optimization method, Default: 1.}

\item{verbose}{Conditional on the verbose level, print the message that each optimization (minimization or maximization) for each group was ended, Default: FALSE.}
}
\value{
An object of class \code{RMSTSens}. The object is a data.frame with the following components:
\item{N}{Total number of subjects}
\item{N.exposed}{The number of subjects in exposed group}
\item{N.unexposed}{The number of subjects in unexposed group}
\item{N.event.exposed}{The number of events in exposed group}
\item{N.event.unexposed}{The number of events in unexposed group}
\item{cen.rate}{Total censoring rate}
\item{cen.rate.exposed}{Censoring rate in exposed group}
\item{cen.rate.unexposed}{Censoring rate in unexposed group}
\item{Lambda}{A scalar or vector of sensitivity parameter \eqn{\Lambda} used}
\item{Tau}{User-specific time point \eqn{\tau}, If tau not specified (NULL), use the minimum between the largest observed event time in each groups}
\item{Method}{A optimization method used}
\item{min.exposed}{The minimum of adjusted RMST based on the shifted propensity score for exposed group}
\item{max.exposed}{The maximum of adjusted RMST based on the shifted propensity score for exposed group}
\item{min.unexposed}{The minimum of adjusted RMST based on the shifted propensity score for unexposed group}
\item{max.unexposed}{The maximum of adjusted RMST based on the shifted propensity score for unexposed group}
\item{RMST.diff.min}{Lower bound of point estimate for between-group difference in adjusted RMST based on shifted propensity score}
\item{RMST.diff.max}{Upper bound of point estimate for between-group difference in adjusted RMST based on shifted propensity score}
The results for the \code{RMSTSens} are printed with the \code{\link{print.RMSTSens}} functions.
To generate result plot comparing sensitivity parameters \eqn{\Lambda} with range of adjusted RMST based on shifted propensity score, use the \code{\link{autoplot.RMSTSens}} functions.
}
\description{
Function for sensitivity analysis of unmeasured confounding for restricted mean survival time using propensity score.
}
\details{
There are four possible methods for our sensitivity analysis.
In general survival analysis setting, if the censoring rates in both exposure group are less than 0.7, the approximate optimization method (methods="Approx") is not inferior to the direct optimization method in terms of both bias and computational time.
Otherwise, one may use the direct optimization method (methods="Optim") to perform the sensitivity analysis, although it takes slightly more computational time than the approximate optimization method.
In special settings, to obtain the solution of optimization problem, only need to compute the objective function of linear fractional programming (methods="LP1" or methods="LP2) for at most m (the number of event) candidates of optimization parameters.
To assess detail of methods and special settings, see Lee et al. (2022).
}
\examples{
dat <- gbsg
dat$size2 <- ifelse(dat$size <= 20, 0,
                    ifelse(dat$size > 20 & dat$size <= 50, 1, 2))
dat$age2 <- dat$age/100
dat$er2 <- dat$er/1000

## Estimation of propensity score
denom.fit <- glm(hormon~(age2)^3+(age2)^3*log(age2)+meno+factor(size2)+sqrt(nodes)+er2,
                 data=dat, family=binomial(link="logit"))
dat$Ps <- predict(denom.fit, type="response")

## Between-group difference in adjusted RMST based on shifted propensity score
## Adjusted RMST with tau equal to 5-year
\dontrun{
# Using direct optimization method
results.optim <- RMSTSens(time="rfstime", status="status", exposure="hormon",
                          level.exposed="1", ps="Ps", data=dat, methods="Optim",
                          use.multicore=TRUE, n.core=parallel::detectCores()/2,
                          lambda=1.5, tau=365.25*5, ini.par=1, verbose=FALSE)
results.optim
}

# Using approximate optimization method
results.approx <- RMSTSens(time="rfstime", status="status", exposure="hormon",
                           level.exposed="1", ps="Ps", data=dat, methods="Approx",
                           use.multicore=TRUE, n.core=2,
                           lambda=1.5, tau=365.25*5, ini.par=1, verbose=FALSE)
results.approx

## Adjusted RMST with not specified tau and with multiple lambda
# Using approximate optimization method
results.approx2 <- RMSTSens(time="rfstime", status="status", exposure="hormon",
                            level.exposed="1", ps="Ps", data=dat, methods="Approx",
                            use.multicore=TRUE, n.core=2,
                            lambda=c(1,1.5), tau=365.25*5, ini.par=1, verbose=FALSE)
results.approx2

}
\references{
Bakbergenuly I, Hoaglin DC, Kulinskaya E (2020):
Methods for estimating between-study variance and overall
effect in meta-analysis of odds-ratios.
\emph{Research Synthesis Methods},
DOI: 10.1002/jrsm.1404
}
\seealso{
\code{\link[RMSTSens]{print.RMSTSens}}, \code{\link[RMSTSens]{autoplot.RMSTSens}}, \code{\link[RMSTSens]{RMSTSens.ci}}
}
\author{
Seungjae Lee \email{seungjae2525@gmail.com}
}
\keyword{methods}
